feat!: Vendor base config and update TypeScript ESLint plugin to v7
The governance of eslint-config-standard and -with-typescript is unclear
and causing trouble for this project. This patch pulls in the relevant
code from -with-standard as a first step towards fully owning the
ESLint configuration. A follow-up patch is planned to do the same for
eslint-config-standard.

By doing this, we are able to upgrade `@typescript-eslint/eslint-plugin`
from v6 to v7, which is the current latest version.

Refs: https://github.com/standard/standard/issues/1948
Refs: https://github.com/standard/standard/issues/1957