feat!: Provide ESLint v9 flat configs
This is a major rewrite that migrates to ESLint v9, and exports a Flat
Config instead of the legacy config. The package is also now a pure
ESM package instead of CommonJS.

README.md describes how to utilize the new config format.

BREAKING CHANGE: There will be many changes in the lint results.